Update image sources in details-component and homecomponent views
This commit updates the image sources in the details-component and homecomponent views of a Laravel application. In the details-component view, the source for the main product image and related product images have been changed. In the homecomponent view, the source for product images in the 'New Arrivals' section has been updated. The new image sources use a different naming convention and include .png files.